Optional map phase, optional cleanup, fix a pooling bug

+++ cmd/bulkloader/main.go
@@ -26,14 +26,22 @@ func main() {
 	flag.StringVar(&opt.rdfFiles, "r", "", "Location of rdf files to load (comma separated)")
 	flag.StringVar(&opt.schemaFile, "s", "", "Location of schema file to load")
 	flag.StringVar(&opt.badgerDir, "p", "", "Location of the final Dgraph directory")
-	flag.StringVar(&opt.tmpDir, "tmp", os.TempDir(), "Temp directory used to use for on-disk "+
+	flag.StringVar(&opt.tmpDir, "tmp", "tmp", "Temp directory used to use for on-disk "+
 		"scratch space. Requires free space proportional to the size of the RDF file.")
 	flag.IntVar(&opt.numGoroutines, "j", runtime.NumCPU(),
 		"Number of worker threads to use (defaults to one less than logical CPUs)")
 	flag.Int64Var(&opt.mapBufSize, "mapoutput_mb", 0,
 		"The estimated size of each map file output. This directly affects the memory usage.")
 	httpAddr := flag.String("http", "localhost:8080", "Address to serve http (pprof)")
+	skipMapPhase := flag.Bool("skip_map_phase", false,
+		"Skip the map phase (assumes that map output files already exist")
+	cleanUpTmp := flag.Bool("cleanup_tmp", true,
+		"Clean up the tmp directory after the loader finishes")
+
 	flag.Parse()
+	if len(flag.Args()) != 0 {
+		log.Fatal("No free args allowed, but got:", flag.Args())
+	}
 
 	x.AssertTruef(opt.mapBufSize > 0, "Please specify how much memory is available for this program.")
 	opt.mapBufSize = opt.mapBufSize << 20 // Convert from MB to B.
@@ -45,12 +53,18 @@ func main() {
 	x.Check(os.MkdirAll(opt.badgerDir, 0700))
 
 	// Create a directory just for bulk loader's usage.
-	os.RemoveAll(opt.tmpDir)
-	x.Check(os.MkdirAll(opt.tmpDir, 0700))
-	defer os.RemoveAll(opt.tmpDir)
+	if !*skipMapPhase {
+		os.RemoveAll(opt.tmpDir)
+		x.Check(os.MkdirAll(opt.tmpDir, 0700))
+	}
+	if *cleanUpTmp {
+		defer os.RemoveAll(opt.tmpDir)
+	}
 
 	loader := newLoader(opt)
-	loader.mapStage()
+	if !*skipMapPhase {
+		loader.mapStage()
+	}
 	loader.reduceStage()
 	loader.writeSchema()
 	loader.writeLease()
